Make the warning policy portable, and clear the 23 warnings it was hiding
The first CI run failed the warning gate, and it was right to. Two separate
faults, both mine, both from the same commit that extended the policy to
every target.

First: -Wno-nontrivial-memcall and -Wno-unnecessary-virtual-specifier
postdate the clang in the CI runner's Xcode. An unknown -Wno- flag is not
ignored - clang answers with -Wunknown-warning-option once per translation
unit, and that warning carries no file path, so it survives a filter written
in terms of paths and looks exactly like a warning in our own code. A policy
meant to keep the build quiet became the thing making the noise.

Each suppression is now probed with check_cxx_compiler_flag and added only
if the compiler knows it. That costs nothing: a compiler that has never
heard of -Wnontrivial-memcall will not emit it either. The core warnings are
old enough to be universal and stay unconditional.

Second, and worse: I claimed this project built warning-clean. It did not. I
checked with an incremental build against an already-built tree, so nothing
recompiled and nothing warned. A genuine build from scratch had 23 warnings
- 16 in tests, 4 in tools, 3 in products - which CI, having no stale tree to
hide behind, found immediately. The claim in the 0.7.0 notes was wrong when
I wrote it.

All 23 are now fixed, each site read rather than blind-cast. One was not a
style nit:

  noteOffWorstSample was an int assigned s.globalSampleBase + n, where
  globalSampleBase is a running long long sample counter - so it truncated
  after about 12 hours of continuous rendering at 48 kHz. The sibling field
  recording the same quantity, worstTransientSample, was already long long.
  Widened to match, and its printf moved from %d to %lld. Casting would have
  made the warning go away and left the truncation.

The rest are explicit casts at narrowing conversions that were already
happening implicitly: size_t/int into double or float in reporting
arithmetic, and int loop counters subscripting vectors, spelled the way the
surrounding lines already spell them.

CI's warning report now separates the two kinds, because they look alike and
have nothing to do with each other: a source warning carries a file:line:col
prefix, a command-line warning does not and means the flags are wrong for
this compiler rather than the code. The job also logs clang --version before
configuring, since a warning that appears only in CI is nearly always a
compiler-version difference.

Verified against a build from scratch this time - 1489 targets, no stale
tree: the gate reports no warnings from PX3 sources, and PX3Tests 1342
passed / 0 failed, PX3Diag regress 0 failures, rtsafety 0 allocations per
block, smoke test audible at every rate and block size.

+++ tests/Tests/TestsEnvelopeModesAndMacros.cpp
@@ -2099,17 +2099,18 @@ void testMacroSystem()
         juce::StringArray ids;
         for (int macro = 0; macro < PX3SynthAudioProcessor::kMacroCount; ++macro)
         {
-            const auto id = destinations[macro]->getParameterID();
+            const auto id = destinations[static_cast<std::size_t>(macro)]->getParameterID();
             ids.add(id);
 
             // Start each base low, so there is room above it to measure into.
-            destinations[macro]->setValueNotifyingHost(0.1f);
+            destinations[static_cast<std::size_t>(macro)]->setValueNotifyingHost(0.1f);
             processor.toggleMacroDestination(macro, id);
         }
 
         const auto effective = [&](int index)
         {
-            return processor.getModulatedNormalisedValue(*destinations[index]);
+            return processor.getModulatedNormalisedValue(
+                *destinations[static_cast<std::size_t>(index)]);
         };
 
         // Each destination belongs to exactly one macro.
@@ -2255,8 +2256,9 @@ void testMacroSystem()
 
                 for (int macro = 0; macro < PX3SynthAudioProcessor::kMacroCount; ++macro)
                 {
-                    targets[macro]->setValueNotifyingHost(0.1f);
-                    knobProcessor.toggleMacroDestination(macro, targets[macro]->getParameterID());
+                    targets[static_cast<std::size_t>(macro)]->setValueNotifyingHost(0.1f);
+                    knobProcessor.toggleMacroDestination(
+                        macro, targets[static_cast<std::size_t>(macro)]->getParameterID());
                 }
 
                 juce::StringArray knobFaults;
@@ -2321,7 +2323,8 @@ void testMacroSystem()
 
             for (int macro = 0; macro < PX3SynthAudioProcessor::kMacroCount; ++macro)
             {
-                saver.toggleMacroDestination(macro, targets[macro]->getParameterID());
+                saver.toggleMacroDestination(
+                    macro, targets[static_cast<std::size_t>(macro)]->getParameterID());
                 saver.getMacroParam(macro).setValueNotifyingHost(0.2f * static_cast<float>(macro + 1));
             }
 
@@ -2336,7 +2339,7 @@ void testMacroSystem()
             auto correct = true;
             for (int macro = 0; macro < PX3SynthAudioProcessor::kMacroCount; ++macro)
             {
-                const auto id = targets[macro]->getParameterID();
+                const auto id = targets[static_cast<std::size_t>(macro)]->getParameterID();
                 const auto mask = loader.getMacroMaskForParameter(id);
                 const auto value = loader.getMacroParam(macro).get();
